requirements.txt
absl-py==1.4.0
asgiref==3.7.2
attrs==22.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.9
djangorestframework==3.15.2
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
flatbuffers==23.3.3
fonttools==4.38.0
greenlet==2.0.2
httpie==3.2.2
idna==3.6
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mediapipe==0.9.1.0
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.24.1
opencv-contrib-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
Pygments==2.18.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
razorpay==1.4.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
SQLAlchemy==2.0.4
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.4
urllib3==2.2.1
Werkzeug==2.2.3
zipp==3.15.0
